Added authentication for postcontroller and added edit post

+++ Forum/Controllers/PostController.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Forum.Models;
+using Microsoft.AspNetCore.Http;
 
 // For more information on enabling MVC for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
@@ -22,9 +23,14 @@ namespace Forum.Controllers
             List<TblPosts> post = TblPosts.GetPostsMatchingString(input);
             return View(post);
         }
+
         [HttpPost]
         public IActionResult Delete(int id)
         {
+            if(HttpContext.Session.GetInt32("ID") != TblPosts.GetPostFromID(id).PoAuthor)
+            {
+                return RedirectToAction("Login", "User");
+            }
             string message = "";
             TblPosts.DeletePostFromID(id, out message);
             ViewBag.message = message;
@@ -34,6 +40,10 @@ namespace Forum.Controllers
         [HttpGet]
         public IActionResult Create(string id)
         {
+            if (HttpContext.Session.GetInt32("ID") == null)
+            {
+                return RedirectToAction("Login", "User");
+            }
             int category = TblCategories.GetCategory(id);
             ViewBag.categoryID = category;
             ViewBag.category = id;
@@ -43,6 +53,11 @@ namespace Forum.Controllers
         [HttpPost]
         public IActionResult Create(TblPosts post)
         {
+            if (HttpContext.Session.GetInt32("ID") == null)
+            {
+                return RedirectToAction("Login", "User");
+            }
+            post.PoAuthor = int.Parse(HttpContext.Session.GetInt32("ID").ToString());
             string message = "";
             TblPosts.InsertPost(post, out message);
             ViewBag.message = message;
@@ -62,10 +77,29 @@ namespace Forum.Controllers
             ViewBag.users = users;
             return View(post);
         }
-        public IActionResult Edit()
-        {
 
-            return View();
+        [HttpGet]
+        public IActionResult Edit(int id)
+        {
+            TblPosts post = TblPosts.GetPostFromID(id);
+            if(post.PoName == null || post.PoAuthor != HttpContext.Session.GetInt32("ID"))
+            {
+                return RedirectToAction("index", "Home");
+            }
+            return View(post);
+        }
+
+        [HttpPost]
+        public IActionResult Edit(TblPosts post)
+        {
+            if (post.PoName == null || post.PoAuthor != HttpContext.Session.GetInt32("ID"))
+            {
+                return RedirectToAction("index", "Home");
+            }
+            string errorMessage = "";
+            TblPosts.EditPost(post.PoId, post.PoName, post.PoContent, out errorMessage);
+            ViewBag.message = errorMessage;
+            return RedirectToAction("Read/" + post.PoId, "Post");
         }
     }
 }
